+++ RELEASE_NOTES.md
@@ -1,1 +1,5 @@
 # _Sharly Chess_ release notes
+
+## Pairing
+
+- Added support for individual and team knockout pairings (single/double elimination, and two-game-match variants) (5.1.0)